\name{bryophytes}
\alias{bryophytes}
\docType{data}
\title{
List of bryophytes
}
\description{
\code{bryophytes} contains species names for 100 Mediterranean bryophytes
}
\usage{data(bryophytes)}
\format{
  A data frame with 100 observations on the following 5 variables.
  \describe{
    \item{\code{Full.name}}{a character vector}
    \item{\code{Genus}}{a character vector}
    \item{\code{Species}}{a character vector}
    \item{\code{Var}}{a character vector}
    \item{\code{Intraspecific}}{a character vector}
  }
}
\examples{
data(bryophytes)
str(bryophytes)
}
\keyword{datasets}
